proper way of handling time_elapsed == 0 and added refreshing of switch status
